requirements.txt
graphene-django
django-filter
gql[requests]
django-crontab